Fix buffer pool serialization buffer lifetime issue (#86)
* Add test showing concurrent serialization bug

* Return copy of serialized data

Allow concurrent serializations to occur by returning a copy of the
buffer holding the data. Golangs bytes.Buffer.Bytes call returns a
pointer to the internal data of the buffer and when the buffer is
modified by calls to Reset or Write the pointer that was returned
earlier will have it's data mutated. This causes issues if Close() is
called on the serialization writer prior to the returned buffer being
used.

* Don't use experimental package

* Add changelog entry